% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_transform.R
\name{drop_underage_group}
\alias{drop_underage_group}
\title{Drop respondents aged 12–17 due to low target case counts}
\usage{
drop_underage_group(df)
}
\arguments{
\item{df}{Decoded dataset}
}
\value{
Filtered dataset excluding "12-17 years"
}
\description{
Drop respondents aged 12–17 due to low target case counts
}
