% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_transform.R
\name{final_transform}
\alias{final_transform}
\title{Final transformations for sparse levels and missing category cleanup}
\usage{
final_transform(df)
}
\arguments{
\item{df}{Dataframe after initial transformation}
}
\value{
Final cleaned dataframe
}
\description{
Final transformations for sparse levels and missing category cleanup
}
