% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_transform.R
\name{load_cchs_data}
\alias{load_cchs_data}
\title{Load and Subset CCHS Data}
\usage{
load_cchs_data(file_path)
}
\arguments{
\item{file_path}{Path to the CCHS dataset (PUMF) CSV file.}
}
\value{
A data.frame with selected and renamed columns.
}
\description{
Loads the raw CCHS CSV file, selects relevant columns, and renames them for readability.
}
