% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_top_features.R
\name{run_gam_on_top_features}
\alias{run_gam_on_top_features}
\title{Run GAM using top 7 categorical features (pre-selected) with random effect smooths}
\usage{
run_gam_on_top_features(df, target, output_dir = "model_results_gam")
}
\arguments{
\item{df}{Modeling dataset (pre-filtered)}

\item{target}{Target variable name (snake_case)}

\item{output_dir}{Directory to save outputs (default = "model_results_gam")}
}
\description{
Run GAM using top 7 categorical features (pre-selected) with random effect smooths
}
