% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_final_logreg.R
\name{train_and_save_model}
\alias{train_and_save_model}
\title{Train and save final logistic regression model (L2 regularized with glmnet)}
\usage{
train_and_save_model(
  df,
  target_variable,
  model_path,
  test_csv_path,
  recipe_path = NULL
)
}
\arguments{
\item{df}{A pre-loaded dataframe}

\item{target_variable}{Target column name (string)}

\item{model_path}{Output path to save trained model as .rds}

\item{test_csv_path}{Output path to save test set}

\item{recipe_path}{Optional: path to save recipe used}
}
\description{
Train and save final logistic regression model (L2 regularized with glmnet)
}
