% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_categorical.R
\name{transform_categorical}
\alias{transform_categorical}
\title{Apply transformation logic to clean and merge categorical levels}
\usage{
transform_categorical(df)
}
\arguments{
\item{df}{Decoded CCHS dataset}
}
\value{
Transformed dataframe
}
\description{
Apply transformation logic to clean and merge categorical levels
}
