% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_mappings.R
\docType{data}
\name{value_mappings}
\alias{value_mappings}
\title{Value mappings for decoding categorical variables in CCHS}
\format{
An object of class \code{list} of length 32.
}
\usage{
value_mappings
}
\description{
Value mappings for decoding categorical variables in CCHS
}
\keyword{datasets}
